refactor: extract static provider builders
